fix(eh_error): 修复 err_on 转换为指针时的类型溢出问题
- 在 eh_error.h 文件中，修改 eh_error_to_ptr 宏定义
- 增加 long 类型转换，转换为指针时正确的保留符号

+++ src/include/eh_error.h
@@ -47,7 +47,7 @@ static inline __attribute_const__ int eh_ptr_to_error(void *ptr){
     return (int)(long)(ptr);
 }
 
-#define eh_error_to_ptr(err_on)             ((void*)(err_on))
+#define eh_error_to_ptr(err_on)             ((void*)(long)(err_on))
 
 #ifdef __cplusplus
 #if __cplusplus
